+++ index.html
@@ -164,6 +164,6 @@
       </div>
     </section>
   </template>
-<script type="module" src="./main.js"></script>
+<script type="module" src="./js/main.js"></script>
 </body>
 </html>
