+++ js/utils.js
@@ -12,4 +12,3 @@ function checkStringLength (string, length) {
   return string.length <= length;
 }
 export {getRandomPositiveInteger}
-
